SimpleNotes/requirements.txt
Django==5.1
asgiref==3.8.1
bcrypt==4.2.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1
pip==24.2
sqlparse==0.5.1
tzdata==2024.1